rename to prompts-api.service
